Adjust python-dateutil version for python < 3.8
CentOS 8 ships with Python 3.6, and the version of the python-dateutil
package in EPEL is at 2.6.1. The datetime parser part of dateutil v2.6.1
is incompatable with Python 3.8 and above, due to a deprecated import
path:

 >   /home/sablosky/.local/share/virtualenvs/amieclient-reJ1Ux08/lib/python3.7/site-packages/dateutil/parser.py:587: DeprecationWarning: Using or importing the ABCs from 'collections' instead of from 'collections.abc' is deprecated, and in 3.8 it will stop working

So for Python 3.7 and below, we install the old version of
python-dateutil, and for 3.8 and above we install v2.8.1.